code style cleanups across the whole source code

+++ libraries/kiwi/kiwi-triplestore/src/main/java/org/apache/marmotta/kiwi/model/rdf/KiWiDoubleLiteral.java
@@ -86,7 +86,7 @@ public class KiWiDoubleLiteral extends KiWiStringLiteral {
      */
     @Override
     public double doubleValue() {
-        return getDoubleContent().doubleValue();
+        return getDoubleContent();
     }
