+++ libraries/kiwi/kiwi-triplestore/src/main/java/org/apache/marmotta/kiwi/model/rdf/KiWiIntLiteral.java
@@ -108,6 +108,6 @@ public class KiWiIntLiteral extends KiWiDoubleLiteral {
      */
     @Override
     public long longValue() {
-        return getIntContent().longValue();
+        return getIntContent();
     }
 }
